reenable HDMI mirroring on GNexus portrait dock
Found by cynewulf @ http://rootzwiki.com/topic/29281-howto-reenable-hdmi-mirroring-on-gnexus-portrait-dock-for-android-403-and-up/

One of the annoyances with the Galaxy Nexus portrait dock is that when used, it does not enable full HDMI mirroring, instead it puts the phone into "dock mode" and only uses the HDMI display under very limited cases, like Youtube videos.

However a simple MHL adapter cable puts the phone into a full mirroring mode, where the phones screen is replicated on the HDMI display. So clearly the Galaxy Nexus is capable of supporting full HDMI mirroring.

Turns out that full mirroring with the portrait dock actually worked under Android 4.0.2, however something happened and 4.0.3 and above (including JB) changed the behavior to "dock" mode.

I figured out what commit for 4.0.3 was responsible for the change in behavior and how to undo it. And now I have HDMI mirroring using the portrait dock working great under JB. Here's the details.

The commit is a change against the Linux kernel and is:

ARM: tuna: support Samsung's MHL adopter ID
Change-Id: Ie105ab9f91589069aa2a658a0ad3c2d798654f70
Signed-off-by: Erik Gilling <konkers@android.com>